+++ mycodo/inputs/ds18b20.py
@@ -70,6 +70,11 @@ class DS18B20Sensor(AbstractInput):
         """
         try:
             self._temperature = self.get_measurement()
+
+            # 85 degrees indicates a communication issue the sensor
+            if self._temperature is 85:
+                return 1
+
             if self._temperature is not None:
                 return  # success - no errors
         except Exception as e: